organize pdf generator code
